Added xml_pretty option
Usage:
  Add xml_pretty to the YAWF::init options, 0 (default) disables
  pretty printing, any other value indicates the number of spaces
  used for indentation.

This is option is meant to be used for debugging only, as a means
to get a good view of what your code actually outputs. Using this
option is discouraged for anything else as the use of spaces and
newlines around HTML tags can give rather annoying results in some
cases.