Removed some PostgreSQL-specific code from YAWF::DB
The YAWF::DB functions should now be (mostly) database-neutral. The only
thing that isn't DB-neutral is the method of getting unicode to work.
Only (recent versions of) DBD::mysql, DBD::sqlite and DBD::Pg are
currently supported.

A later version of DBI might have a better defined interface to do this,
let's hope we'll see some new developments in that aspect in the near
future.