TUWF::Request: Implemented own parameter parsing and handling
This removes the dependency on CGI::Minimal.

Also added reqGET() and reqPOST() methods, which can be used instead of
reqParam().

multipart/form-data is not supported yet, and there are a few other
TODO's left.

The behaviour of reqParam() changed a little with regards to fetching
parameters that are not present in the request: previously it would
return an empty string, it will now return undef.