TUWF.pm: Check the return value of eval()'s instead of $@
A very old "bug". It is possible to make eval() fail without having $@
set. I believe the trick to do so was the following code:

  {
    local $!; # or $@
    die "This message is ignored.";
  }

Checking for the return value of eval() is more reliable, as that is
guaranteed to return undef if something went wrong, and the 'return'
value otherwise, which we can influence ourself by making all eval()s
return the number 1.