TUWF::XML: Renamed xml_escape to html_escape and fixed xml_escape
xml_escape(), as the name suggests, escapes XML data, which would
exclude converting \n to <br /> tags.

html_escape() is the same as the previous xml_escape(), as that
behaviour *is* quite useful in certain cases.

This change also affects all code that assumes tag() values are
html_escape()'d. If it turns out that VNDB had a lot of instances where
I did make that assumption, I'll probably add a
reference-to-scalar-uses-html_escape thing... But it may be nice to
reserve the scalarref for other use, like for lit().